changes the forum show to the id #51
