+++ coveralls.json
@@ -7,6 +7,7 @@
     "test/support/model_case.ex",
     "test/support/channel_case.ex",
     "web/web.ex",
-    "web/views/error_helpers.ex"
+    "web/views/error_helpers.ex",
+    "web/models/forum.ex"
   ]
 }
