+++ web/models/forum.ex
@@ -1,0 +1,9 @@
+defmodule Feedback.Forum do
+  use Feedback.Web, :model
+
+  defimpl Phoenix.Param, for: Feedback.Forum do
+    def to_param(%{id: id}) do
+      "#{id}"
+    end
+  end
+end
