+++ web/templates/forum/forum.html.eex
@@ -1,6 +1,6 @@
 <div class="feedback-dashboard-title title">Feedback Forum</div>
 <%= for feedback <- @public_feedback do %>
-  <a class="link" href="/forum/<%= feedback.permalink_string %>">
+  <a class="link" href="/forum/<%= feedback.id %>">
     <div class="feedback-item">
       <img class="mood-image-forum" src="/images/<%= feedback.mood %>.svg" />
       <%= if feedback.response != nil do %>
